Separate rule provider,context and hook in files
